general refactoring
- move most methods from the trait to a separate service
- fix parts that were using the old constants class
- improve/fix some parts
