+++ src/ConsoleZoo.php
@@ -2,18 +2,10 @@
 
 namespace Deerdama\ConsoleZoo;
 
-use ReflectionClass;
-
 trait ConsoleZoo
 {
-    /** @var array */
-    protected $currentDefaults = [];
-
-    /** @var array */
-    protected $constants = [];
-
-    /** @var array */
-    protected $defaultIcons = [];
+    /** @var OutputService */
+    protected $zooService;
 
     /**
      * set default style (optional)
@@ -22,9 +14,8 @@ trait ConsoleZoo
      */
     public function zooSetDefaults($param)
     {
-        $this->defaultIcons = $param['icons'] ?? [];
-        unset($param['icons']);
-        $this->currentDefaults = $this->prepareParameters($param, true);
+        $this->start();
+        $this->zooService->setDefaults($param);
     }
 
     /**
@@ -40,6 +31,16 @@ trait ConsoleZoo
     }
 
     /**
+     * check/create service instance
+     */
+    public function start()
+    {
+        if (!$this->zooService) {
+            $this->zooService = app(OutputService::class);
+        }
+    }
+
+    /**
      * form the parameters and output message
      *
      * @param string $message
@@ -48,265 +49,9 @@ trait ConsoleZoo
      */
     public function zooOutput($message = "", $icons = [], $param = [], $ignoreDefault = false)
     {
-        $this->getConstants();
-
-        if (!$icons && $this->defaultIcons) {
-            $icons = $this->defaultIcons;
-        }
-
-        $ansi = $this->prepareSequence($param, $ignoreDefault);
-        $icons = $this->prepareIcons($icons);
-        $message = $this->prepareMessage($message, $ansi);
-
-        $this->output->writeln(" " . $icons . $ansi . $message);
-    }
-
-    /**
-     * @param $parameters
-     * @param bool $ignoreDefaults
-     * @return array
-     */
-    private function prepareParameters($parameters, $ignoreDefaults = false)
-    {
-        $this->getConstants();
-        $currentDefaults = $this->currentDefaults;
-        $newParam = [];
-
-        foreach ($parameters as $key => $value) {
-            if (is_string($key)) {
-                $newParam[$key] = $value;
-            } else {
-                if (substr($value, 0, 3) === 'no_') {
-                    unset($currentDefaults[substr($value, 3)]);
-                    unset($parameters[substr($value, 3)]);
-                    unset($newParam[substr($value, 3)]);
-                } else if (isset($this->constants[strtoupper($value)])) {
-                    $newParam[$value] = $this->constants[strtoupper($value)];
-                }
-            }
-        }
-
-        $result = $this->getColors($newParam);
-
-        if ($ignoreDefaults !== true) {
-            $result = array_merge($currentDefaults, $result);
-        }
-
-        return $result;
-    }
-
-    /**
-     * check if the constants were loaded, get them if not
-     */
-    private function getConstants()
-    {
-        if (!count($this->constants)) {
-            $refl = new ReflectionClass(Zoo::class);
-            $this->constants = $refl->getConstants();
-        }
-    }
-
-    /**
-     * get the ansi sequences for color parameters
-     *
-     * @param array $parameters
-     * @return array
-     */
-    private function getColors($parameters)
-    {
-        if (isset($parameters['color'])) {
-            $color = $this->findColor($parameters['color']);
-
-            if (is_array($color)) {
-                $parameters['color'] = Zoo::SET_COLOR_RGB . implode(';', $color);
-            } else if (is_integer($color)) {
-                $parameters['color'] = Zoo::SET_COLOR . $color;
-            } else {
-                unset($parameters['color']);
-            }
-        }
-
-        if (isset($parameters['background'])) {
-            $color = $this->findColor($parameters['background']);
-
-            if (is_array($color)) {
-                $parameters['background'] = Zoo::SET_BG_RGB . implode(';', $color);
-            } else if (is_integer($color)) {
-                $parameters['background'] = Zoo::SET_BG . $color;
-            } else {
-                unset($parameters['background']);
-            }
-        }
-
-        return $parameters;
-    }
-
-    /**
-     * @param $color
-     * @return int|array|null
-     */
-    private function findColor($color)
-    {
-        if (is_string($color)) {
-            $color = str_replace(' ', '_', $color);
-            $color = strpos(strtoupper($color), 'COLOR_') === 0 ? $color : 'COLOR_' . $color;
-            $color = isset($this->constants[strtoupper($color)]) ? $this->constants[strtoupper($color)] : null;
-        }
-
-        return $color;
-    }
-
-    /**
-     * prepare ansi escape sequence for effects and style
-     *
-     * @param array $param
-     * @return string
-     */
-    private function prepareSequence($param = [], $ignoreDefault = false): string
-    {
-        if (!count($param) && !count($this->currentDefaults)) {
-            return "";
-        }
-
-        $sequenceParam = $this->prepareParameters($param, $ignoreDefault);
-
-        if (!count($sequenceParam)) {
-            return "";
-        }
-
-        $ansi = "\e[" . implode(';', $sequenceParam);
-        $ansi = rtrim($ansi, ';') . 'm';
-
-        return $ansi;
-    }
-
-    /**
-     * @param array|string $icons
-     * @return string
-     */
-    private function prepareIcons($icons): string
-    {
-        if (!$icons || !count((array)$icons)) {
-            return "";
-        }
-
-        if (is_array($icons)) {
-            foreach ($icons as $icon) {
-                $result[] = $this->getIconCode($icon);
-            }
-        } else {
-            $result[] = $this->getIconCode($icons);
-        }
-
-        $text = implode(' ', $result) . ' ';
-
-        return $text;
-    }
-
-    /**
-     * @param string $icon
-     * @return string
-     */
-    private function getIconCode($icon)
-    {
-        if (is_array($icon) && isset($icon['utf8'])) {
-            $icon = $icon['utf8'];
-        }
-
-        if (is_string($icon) && isset($this->constants[strtoupper($icon)])) {
-            $icon = $this->constants[strtoupper($icon)]['utf8'];
-        }
-
-        return $icon;
-    }
-
-    /**
-     * check if the message contains inline parameters and return the final text
-     *
-     * @param string $message
-     * @param string $mainSequence
-     * @return string
-     */
-    private function prepareMessage($message, $mainSequence = "")
-    {
-        $message = $this->parseInlineParam($message, $mainSequence);
-        $message = $this->parseInlineIcons($message);
-        $result = $message . Zoo::RESET;
-
-        return $result;
-    }
-
-    /**
-     * check if the message contains inline parameters
-     *
-     * @param string $message
-     * @param string $origSequence
-     * @return mixed|string
-     */
-    private function parseInlineParam($message, $origSequence)
-    {
-        preg_match('/<zoo.*<\/zoo>/U', $message, $match);
-
-        if ($match) {
-            $inline = $match[0];
-            preg_match('/(?<=>).*(?=<\/zoo>)/U', $inline, $content);
-            preg_match('/(?<=color=[",\']).*(?=[",\'])/U', $inline, $color);
-            preg_match('/(?<=background=[",\']).*(?=[",\'])/U', $inline, $background);
-
-            if ($color) {
-                preg_match('/(?<=\[).*(?=])/U', $color[0], $colorArr);
-                $param['color'] = $colorArr ? explode(',', str_replace(' ', '', $colorArr[0])) : $color[0];
-                $inline = preg_replace('/color=[",\'].*[",\']/', '', $inline);
-            }
-
-            if ($background) {
-                preg_match('/(?<=\[).*(?=])/U', $background[0], $backgroundArr);
-                $param['background'] = $backgroundArr ? explode(',', str_replace(' ', '', $backgroundArr[0])) : $background[0];
-                $inline = preg_replace('/background=[",\'].*[",\']/U', '', $inline);
-            }
-
-            preg_match('/(?<=<zoo ).*(?=>)/U', $inline, $others);
-
-            if ($others) {
-                $others = explode(' ', $others[0]);
-
-                foreach ($others as $other) {
-                    if (!$other) {
-                        continue; //don't include extra white spaces
-                    }
-                    $param[] = $other;
-                }
-            }
-
-            $sequence = $this->prepareSequence($param, true);
-            $result = $sequence . $content[0] . Zoo::RESET . $origSequence;
-
-            $message = str_replace($match, $result, $message);
-            return $this->parseInlineParam($message, $origSequence);
-        }
-
-        return $message;
-    }
-
-    /**
-     * check if the message contains icons
-     *
-     * @param $message
-     * @return mixed
-     */
-    private function parseInlineIcons($message)
-    {
-        preg_match('/<icon.*<\/icon>/U', $message, $inline);
-
-        if ($inline) {
-            preg_match('/(?<=<icon>).*(?=<\/icon>)/U', $inline[0], $icon);
-            $icon = $icon ? $this->getIconCode($icon[0]) : "";
-            $message = str_replace($inline, $icon, $message);
-
-            return $this->parseInlineIcons($message);
-        }
-
-        return $message;
+        $this->start();
+        $output = $this->zooService->prepareOutput($message, $icons, $param, $ignoreDefault);
+        $this->output->writeln($output);
     }
 
     /**
@@ -315,12 +60,9 @@ trait ConsoleZoo
      */
     public function zooInfo($message = "", $param = [])
     {
-        $parameters = Zoo::infoDefaults($param);
-        $icons = $parameters['icons'];
-        $message = isset($param['icons']) || !is_string($icons) ? $message : $message . "\e[0m " . $icons;
-        unset($parameters['icons']);
-
-        $this->zooOutput($message, $icons, $parameters, true);
+        $this->start();
+        $output = $this->zooService->defaultOutputs($message, $param, 'info');
+        $this->output->writeln($output);
     }
 
     /**
@@ -329,12 +71,9 @@ trait ConsoleZoo
      */
     public function zooSuccess($message = "", $param = [])
     {
-        $parameters = Zoo::successDefaults($param);
-        $icons = $parameters['icons'];
-        $message = isset($param['icons']) || !is_string($icons) ? $message : $message . "\e[0m " . $icons;
-        unset($parameters['icons']);
-
-        $this->zooOutput($message, $icons, $parameters, true);
+        $this->start();
+        $output = $this->zooService->defaultOutputs($message, $param, 'success');
+        $this->output->writeln($output);
     }
 
     /**
@@ -343,12 +82,9 @@ trait ConsoleZoo
      */
     public function zooWarning($message = "", $param = [])
     {
-        $parameters = Zoo::warningDefaults($param);
-        $icons = $parameters['icons'];
-        $message = isset($param['icons']) || !is_string($icons) ? $message : $message . "\e[0m " . $icons;
-        unset($parameters['icons']);
-
-        $this->zooOutput($message, $icons, $parameters, true);
+        $this->start();
+        $output = $this->zooService->defaultOutputs($message, $param, 'warning');
+        $this->output->writeln($output);
     }
 
     /**
@@ -357,12 +93,9 @@ trait ConsoleZoo
      */
     public function zooError($message = "", $param = [])
     {
-        $parameters = Zoo::errorDefaults($param);
-        $icons = $parameters['icons'];
-        $message = isset($param['icons']) || !is_string($icons) ? $message : $message . "\e[0m " . $icons;
-        unset($parameters['icons']);
-
-        $this->zooOutput($message, $icons, $parameters, true);
+        $this->start();
+        $output = $this->zooService->defaultOutputs($message, $param, 'error');
+        $this->output->writeln($output);
     }
 
     /**
@@ -373,38 +106,9 @@ trait ConsoleZoo
      */
     public function surprise($message = "", $param = [])
     {
-        $this->getConstants();
-        $icon = null;
-
-        if (!isset($param['category']) || !in_array($param['category'], Zoo::CATEGORIES)) {
-            $param['category'] = null;
-        }
-
-        while (!$icon) {
-            $rand = $this->constants[array_rand($this->constants)];
-
-            if (isset($rand['type']) && $rand['type'] === 'icon'
-                && (!$param['category'] || $param['category'] === $rand['category'])) {
-
-                $icon = $rand['utf8'];
-            }
-        }
-
-        if (!isset($param['color']) && !isset($this->currentDefaults['color'])) {
-            while (!isset($param['color'])) {
-                $rand = array_rand($this->constants);
-
-                if (strpos($rand, 'COLOR_') === 0) {
-                    $param['color'] = $rand;
-                }
-            }
-        }
-
-        unset($param['category']);
-        unset($param['icons']);
-        $message = $message . "\e[0m " . $icon;
-
-        $this->zooOutput($message, $icon, $param);
+        $this->start();
+        $output = $this->zooService->surpriseOutput($message, $param);
+        $this->output->writeln($output);
     }
 }
