+++ src/Others.php
@@ -5,6 +5,11 @@ namespace Deerdama\ConsoleZoo;
 class Others
 {
     /** styles and effects **/
+    const SET_COLOR = "38;5;";
+    const SET_COLOR_RGB = "38;2;";
+    const SET_BG = "48;5;";
+    const SET_BG_RGB = "48;2;";
+
     const RESET = "\e[0m";
     const BOLD = 1;
     const FAINT = 2;
